- [addeded] check type limits on set value - [addeded] check type on set value - [addeded] option Bounds for integer and floating point - [addeded] option Default for bool, integer, floating point and string - [addeded] option NotEmpty for string - [addeded] Visual C++ 2019 and Microsoft (R) C/C++ Optimizing Compiler Version 19.26.28806 for x64 | Windows 64-bit (except tests)
- [changed] exception messages

- [removed] unsigned long

+++ example/containers/containers.cpp
@@ -2,6 +2,7 @@
 #include <list>
 #include <map>
 #include <sstream>
+#include <string>
 #include <vector>
 
 #include "struct_mapping/struct_mapping.h"
@@ -41,7 +42,7 @@ struct Star {
 struct PlanetSystem {
 	std::string name;
 	long long age;
-	long major_axis;
+	long long major_axis;
 	bool populated;
 	Star star;
 	std::list<Planet> planets;
